pyproject.toml
[project]
name = "nectarlite"
version = "0.1.0"
description = "Lean Mean Hive Engine Machine"
readme = "README.md"
authors = [{ name = "Michael Garcia", email = "thecrazygm@gmail.com" }]
requires-python = ">=3.14"
dependencies = ["requests", "pycryptodome", "scrypt", "ecdsa>=0.19.1"]

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["pytest>=8.4.2"]
